+++ Sources/SendBirdSDK.xcframework/ios-arm64_armv7/SendBirdSDK.framework/Headers/SBDBaseMessageParams.h
@@ -76,7 +76,7 @@
 /// The unique ID of a parent message. A parent message is a message that has a thread of replies. If the message sent through the [`sendUserMessageWithParams:completionHandler:`](../Classes/SBDBaseChannel.html#//api/name/sendUserMessageWithParams:completionHandler:) or [`sendFileMessageWithParams:completionHandler:`](../Classes/SBDBaseChannel.html#//api/name/sendFileMessageWithParams:completionHandler:) method is a parent message, the value of this property is 0. If the message is a reply to a parent message, the value is the message ID of the parent message.
 /// @note The default value is `0`.
 /// @since 3.0.181
-@property (atomic) long long parentMessageId;
+@property (atomic, assign) long long parentMessageId;
 
 /// Push notification options for the Apple critical alert.
 /// The default value is `nil`.