+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDMain.h
@@ -25,6 +25,7 @@
 #import "SBDUserListQuery.h"
 
 @class SBDDatabase;
+@class SBDPollListQuery, SBDPollVoterListQuery;
 
 typedef void(^SBDBackgroundSessionBlock)(void);