+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDMessageChangeLogsParams.h
@@ -40,6 +40,11 @@
 /// @since 3.0.181
 @property (atomic) BOOL includeThreadInfo;
 
+/// Determines whether to include poll information of the message in the results
+/// @note The default value is `NO`
+/// @since 3.0.232
+@property (atomic) BOOL includePollDetails;
+
 /// Create `SBDPreviousMessageListQuery` from a given SBDPreviousMessageListQuery instance.
 /// @param query previous message list query
 /// @return A created instance of `SBDMessageChangeLogsParams`